refactor: use ADR architecture for wedding pages
I felt like the code was starting to get messy. And after thinking about
it decided it was because I was mixing concerns in namespaces. The view
namespace was making database requests, for example, in addition to
rendering HTML responses and parsing request parameters.

I considered switching to an MVC architecture, but decided that the
ADR (Action-Domain-Responder) architecture [1] is more appropriate for a
server-side web application.

This commit changes only the /wedding/* pages. I'm happy with the
results though. So I'll convert the rest of the site next.

[1]: http://pmjones.io/adr/