refactor: rename "plus-ones" to "party-size"
The data in the "plus_ones" column in the database and used throughout
the code now represents the total size of the party. This commit changes
the code to reflect that it's a party size instead of plus ones. It also
fixes the calculation used to figure out the number of attendees going
to a reception.